Reduce code duplication in pty impls
Fixes #203663
