+++ src/vs/workbench/contrib/terminal/common/basePty.ts
@@ -1,0 +1,112 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Emitter } from 'vs/base/common/event';
+import { Disposable } from 'vs/base/common/lifecycle';
+import { mark } from 'vs/base/common/performance';
+import { URI } from 'vs/base/common/uri';
+import type { IPtyHostProcessReplayEvent, ISerializedCommandDetectionCapability } from 'vs/platform/terminal/common/capabilities/capabilities';
+import { ProcessPropertyType, type IProcessDataEvent, type IProcessProperty, type IProcessPropertyMap, type IProcessReadyEvent, type ITerminalChildProcess } from 'vs/platform/terminal/common/terminal';
+
+/**
+ * Responsible for establishing and maintaining a connection with an existing terminal process
+ * created on the local pty host.
+ */
+export abstract class BasePty extends Disposable implements Partial<ITerminalChildProcess> {
+	protected readonly _properties: IProcessPropertyMap = {
+		cwd: '',
+		initialCwd: '',
+		fixedDimensions: { cols: undefined, rows: undefined },
+		title: '',
+		shellType: undefined,
+		hasChildProcesses: true,
+		resolvedShellLaunchConfig: {},
+		overrideDimensions: undefined,
+		failedShellIntegrationActivation: false,
+		usedShellIntegrationInjection: undefined
+	};
+	protected readonly _lastDimensions: { cols: number; rows: number } = { cols: -1, rows: -1 };
+	protected _inReplay = false;
+
+	protected readonly _onProcessData = this._register(new Emitter<IProcessDataEvent | string>());
+	readonly onProcessData = this._onProcessData.event;
+	protected readonly _onProcessReplayComplete = this._register(new Emitter<void>());
+	readonly onProcessReplayComplete = this._onProcessReplayComplete.event;
+	protected readonly _onProcessReady = this._register(new Emitter<IProcessReadyEvent>());
+	readonly onProcessReady = this._onProcessReady.event;
+	protected readonly _onDidChangeProperty = this._register(new Emitter<IProcessProperty<any>>());
+	readonly onDidChangeProperty = this._onDidChangeProperty.event;
+	protected readonly _onProcessExit = this._register(new Emitter<number | undefined>());
+	readonly onProcessExit = this._onProcessExit.event;
+	protected readonly _onRestoreCommands = this._register(new Emitter<ISerializedCommandDetectionCapability>());
+	readonly onRestoreCommands = this._onRestoreCommands.event;
+
+	constructor(
+		readonly id: number,
+		readonly shouldPersist: boolean
+	) {
+		super();
+	}
+
+	async getInitialCwd(): Promise<string> {
+		return this._properties.initialCwd;
+	}
+
+	async getCwd(): Promise<string> {
+		return this._properties.cwd || this._properties.initialCwd;
+	}
+
+	handleData(e: string | IProcessDataEvent) {
+		this._onProcessData.fire(e);
+	}
+	handleExit(e: number | undefined) {
+		this._onProcessExit.fire(e);
+	}
+	handleReady(e: IProcessReadyEvent) {
+		this._onProcessReady.fire(e);
+	}
+	handleDidChangeProperty({ type, value }: IProcessProperty<any>) {
+		switch (type) {
+			case ProcessPropertyType.Cwd:
+				this._properties.cwd = value;
+				break;
+			case ProcessPropertyType.InitialCwd:
+				this._properties.initialCwd = value;
+				break;
+			case ProcessPropertyType.ResolvedShellLaunchConfig:
+				if (value.cwd && typeof value.cwd !== 'string') {
+					value.cwd = URI.revive(value.cwd);
+				}
+		}
+		this._onDidChangeProperty.fire({ type, value });
+	}
+	async handleReplay(e: IPtyHostProcessReplayEvent) {
+		mark(`code/terminal/willHandleReplay/${this.id}`);
+		try {
+			this._inReplay = true;
+			for (const innerEvent of e.events) {
+				if (innerEvent.cols !== 0 || innerEvent.rows !== 0) {
+					// never override with 0x0 as that is a marker for an unknown initial size
+					this._onDidChangeProperty.fire({ type: ProcessPropertyType.OverrideDimensions, value: { cols: innerEvent.cols, rows: innerEvent.rows, forceExactSize: true } });
+				}
+				const e: IProcessDataEvent = { data: innerEvent.data, trackCommit: true };
+				this._onProcessData.fire(e);
+				await e.writePromise;
+			}
+		} finally {
+			this._inReplay = false;
+		}
+
+		if (e.commands) {
+			this._onRestoreCommands.fire(e.commands);
+		}
+
+		// remove size override
+		this._onDidChangeProperty.fire({ type: ProcessPropertyType.OverrideDimensions, value: undefined });
+
+		mark(`code/terminal/didHandleReplay/${this.id}`);
+		this._onProcessReplayComplete.fire();
+	}
+}
